Some Micro Changes
Changes include Camera size increased (makes everything look smaller) and ship speeds increased but other things too. I can push to develop whenever cause Im the only person working on this so it's fine. Also, since I am the only person working on this, why am I writing this description? Perhaps for myself in the future? Nah, I never read these anyways. Maybe I do it to fill the void in my soul and make it seem like I'm a part of something. No doubt though this will soon be in a place where it might as well not exist.